removed redundancies in pathways
